requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.121.1
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pydantic==2.12.4
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
python==3.9
